refactor(api-e2e): split endpoint-registry fixture by domain
Bean: api-kxth
